Moved LatestFileVersion to Helpers
